+++ src/IECoreRenderMan/Object.cpp
@@ -198,4 +198,6 @@ void Object::link( const IECore::InternedString &type, const IECoreScenePreview:
 
 void Object::assignID( uint32_t id )
 {
+	m_extraAttributes.SetInteger( Rix::k_identifier_id, id );
+	attributes( m_attributes.get() );
 }