+++ Changes.md
@@ -13,6 +13,8 @@ Improvements
 - ShaderAssignment : The `scene:path` context variable is now available in Switches connected directly to the `ShaderAssignment.shader` input. This allows different shaders to be assigned to different locations using a single ShaderAssignment node. Please note that the `scene:path` context variable remains unavailable to the individual shader nodes themselves for performance reasons.
 - 3Delight, Cycles, OpenGL : Added support for custom EXR metadata, using `header:*` parameters on the output definition.
 - RenderManAttributes, RenderManOptions : Plugs now respect minimum and maximum values specified by RenderMan.
+- Cycles : Added support for `layerName` parameter in outputs, to control the naming of channels in EXR outputs.
+- StandardOptions : Added render manifest option.
 
 Fixes
 -----